requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.4.0
flake8==6.1.0
black==24.1.0
